S12 enabling gnss services

+++ utils/GEOnex_Base_ESP32_v2/src/mqtt_manager.cpp
@@ -103,7 +103,7 @@ void publishGPSData(float latitude, float longitude, int satellites, String time
     }
 }
 
-void publishData(String deviceId, String status, float latitude, float longitude, int satellites,
+void publishData(String deviceId, String status, double latitude, double longitude, int satellites,
                  String time, int battery, int wifi)
 
 {
@@ -137,7 +137,7 @@ void publishData(String deviceId, String status, float latitude, float longitude
     }
 }
 
-void publishBaseFix(float latitude, float longitude)
+void publishBaseFix(double latitude, double longitude)
 
 {
     if (!mqttConnected())
